mcp_promt_opt/pyproject.toml
[project]
name = "mcp-promt-opt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.34.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "openai>=1.40.0",
    "pydantic>=2.7.0",
    "pytest>=8.3.0",
    "python-dotenv>=1.0.1",
]
